+++ src/components/Theme/Theme.utilities.ts
@@ -1,0 +1,2 @@
+export const getRootThemeEl = () =>
+	document.querySelector("[data-rs-root]") || document.documentElement;